Improve SelectFilesApp constructor and add debug logging
The SelectFilesApp constructor has been updated to now take in 'template_names' in addition to 'files'. This aims to optimize the process of file checking based on template names. Debug logging has also been added to 'ghignore.rs' to track option parsing and will aid in finding potential errors or bugs in future.